core/requirements.txt
fastapi
uvicorn
google-generativeai
playwright
playwright-stealth
ddgs
python-dotenv
pydantic
